requirements.txt
streamlit>=1.36.0
openai>=1.35.0
boto3>=1.34.0
google-generativeai>=0.7.2
python-dotenv>=1.0.1
